Add backpressure ability
This commit adds a, rather involved, ability to back-pressure
commands. When the client is configured to enforce
throughput using  `maxCommandsInQueue` and `maxCommandsInProcess`
the user code is expected to apply some external throttle.
This kind of mechanism is not a part of the client. it is added here
and can be switched on by adding a truthy `block-on-queue-full?` flag
on client creation.
Doing so will enforce, using a semaphore, that the commands queue
does not get filled up (which causes the client to throw a
`AsyncQueuefull` exception). All futures returned will only be completed after releasing the semaphore.